Testperfil

+++ Codigo/SeguradosAPI/Models/UsuarioHasPerguntaModel.cs
@@ -4,6 +4,6 @@
     {
         public int IdUsuario { get; set; }
         public int IdPergunta { get; set; }
-        public byte Acertou { get; set; }
+        public int Acertou { get; set; }
     }
 }
